// src/enums/WebUrl.ts
enum WebUrl {
  BASE = "/",
  LOGIN = "/login",
  DASHBOARD = "/dashboard",
  ACCOUNTING = "/dashboard/accounting",
  BANKING = "/dashboard/banking",
  REPORTS = "/dashboard/reports",
  CONTACTS = "/dashboard/contacts",
  INVOICING = "/dashboard/invoicing",
  QUERIES = "/dashboard/queries",
}

export default WebUrl;